Code review EP-806: - EP-813 add TODOs on what could be moved into Builders - move "examples" and "scripts" under tvb_epilepsy.top
