+++ tvb_epilepsy/base/constants/configurations.py
@@ -1,9 +1,13 @@
 import os
 import platform
 from datetime import datetime
+import tvb_epilepsy
 
 
 USER_HOME = os.path.expanduser("~")
+TVB_EPILEPSY_PATH = os.path.dirname(tvb_epilepsy.__file__)
+STATS_MODELS_PATH = os.path.join(TVB_EPILEPSY_PATH, "service", "model_inversion", "stan", "models")
+
 RUN_ENV = "local"
 if "RUN_ENV" in os.environ:
     RUN_ENV = os.environ["RUN_ENV"]
@@ -21,13 +25,11 @@ else:
         # DATA_CUSTOM = os.path.join(USER_HOME, 'CBR', 'svn', 'episense', 'demo-data')
         DATA_TVB = os.path.join(USER_HOME, 'CBR', 'svn', 'tvb', 'tvb-data', 'tvb-data')
         DATA_CUSTOM = os.path.join(FOLDER_VEP, 'TVB3')
-        TVB_EPILEPSY_PATH =  "/Users/dionperd/VEPtools/git/tvb-epilepsy/tvb_epilepsy"
     else:
         FOLDER_VEP_TESTS = os.path.join(FOLDER_VEP_ONLINE, 'tests')
         # DATA_CUSTOM = os.path.join(USER_HOME, 'CBR_software', 'svn-episense', 'demo-data')
         DATA_TVB = os.path.join(USER_HOME, 'CBR_software', 'svn-tvb', 'tvb-data', 'tvb-data')
         DATA_CUSTOM = os.path.join(FOLDER_VEP, 'TVB3')
-        TVB_EPILEPSY_PATH = "/Users/dionperd/VEPtools/software/git/tvb-epilepsy/tvb_epilepsy"
     if not (os.path.isdir(FOLDER_VEP_TESTS)):
         os.mkdir(FOLDER_VEP_TESTS)
     # Folder where input data will be
@@ -42,7 +44,7 @@ else:
     FOLDER_FIGURES = os.path.join(FOLDER_VEP_TESTS, 'figures' + datetime.strftime(datetime.now(), '%Y-%m-%d_%H-%M'))
     if not (os.path.isdir(FOLDER_FIGURES)):
         os.mkdir(FOLDER_FIGURES)
-    STATS_MODELS_PATH = os.path.join(VEP_SOFTWARE_PATH, "tvb-epilepsy", "tvb_epilepsy", "stan")
+
 CMDSTAN_PATH = os.path.join(USER_HOME, "ScientificSoftware/git/cmdstan")
 VERY_LARGE_SIZE = (40, 20)
 VERY_LARGE_PROTRAIT = (30, 50)
