+++ tvb_epilepsy/top/examples/test_equations.py
